sched/fair: Reduce balance interval to 0 if we have a misfit task
It is quite a common occurrence that we have a misfit task (a task
running on a CPU which is close to the capacity limit of that CPU)
in the system AND we have recently done a load balance operation.
Normal load-balance rate limiting means we can leave those tasks on
the restrictive CPU for a few hundred ms before we move them.

Inspired by the work Leo Yan did on balancing misfits.

This patch reduces the balance interval to zero when there are such
tasks running.

There appears to be no impact on interactive performance from this
patch.

Change-Id: I71d74734d3bd6bca80476c0f466ba48914b7ac02
Signed-off-by: Chris Redpath <chris.redpath@arm.com>
Signed-off-by: Joe Maples <joe@frap129.org>