simple_lmk: Only initialize driver after LMKD writes minfree settings
When booted in recovery, the OOM ADJ scores are not set according to what
they're expected to be in Android, which causes Simple LMK to misbehave.

In order to ensure that Simple LMK is only used with a compatible
userspace (i.e., Android), don't initialize the driver until LMKD in
Android attempts to write minfree settings. This is an indicator that
LMKD/Android is present and that Simple LMK can therefore be used.

Change-Id: Iebbb0d702145a00cdd033ebe95476fd9789d3a2d
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: RJ Trujillo <certifiedblyndguy@gmail.com>